USER PROFILE: Protect Edit / Bugfix

+++ src/controllers/videoController.js
@@ -15,17 +15,30 @@ const watch = async (req, res) => {
 };
 const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id },
+    } = req.session;
     const video = await Video.findById(id);
     if (!video)
         return res.status(404).render('404', { pageTitle: 'Video not found' });
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect('/');
+    }
     return res.render('edit', { pageTitle: `Edit ${video.title}`, video });
 };
 const postEdit = async (req, res) => {
+    const {
+        user: { _id },
+    } = req.session;
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if (!video)
         return res.status(404).render('404', { pageTitle: 'Video not found' });
+
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect('/');
+    }
     await Video.findByIdAndUpdate(id, {
         title,
         description,
@@ -69,6 +82,13 @@ const postUpload = async (req, res) => {
 
 const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id },
+    } = req.session;
+    const video = await Video.findById(id);
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect('/');
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect('/');
 };
